refactor: separate i18n catalogs from runtime
